Refactor valid image getter
